Tidy up and docstrings
